Revised version checking so that the old versions style can be notified.
- Look for 'versions.txt' rather than 'current.txt' on the web site so
that current.txt can be modified to have a bogus newer version number to
ensure old versions will notice the update (the last released version
with the old format was 2011.01.02.1054, so it has to be larger than
that, yet the new version numbering scheme is more traditional (e.g.
4.0.0) and would be considered smaller).
- No longer mention the exact version that can be updated to. This means
we'll need to re-localized the OUT_OF_DATE string.
- Store the last version we've seen under a new key so that the bogus
value doesn't prevent future updates.