Initial work on GCS-26: Add link to PDFs.
Needs some work to pretty up the preferences. Also needs a mechanism to
set a page offset on a per-PDF basis, as some PDFs do not open to the
correct page (i.e. the page you want is 28, but if you open page 28, it
is actually page 27 because they added a cover page and didn't tell
their PDF generator to place updated page references into the PDF).