GCS-20: Equipment option to nullify weight.
A piece of equipment can now have a feature added to it that reduces the
effective weight of its contained items by either an absolute quantity
or a percentage. To make this work properly, a small change to the way
the extended weight field is calculated and displayed was made. In the
past, it would always show the weight of the container itself plus all
of its contents. Now, it only shows the weight of its contents if those
contents are actually being carried/equipped. This change allows
something like the Pathfinder Bag of Holding Type I to be created by
specifying a weight reduction of 250 lbs as a feature and a maximum
contained weight of 0 lbs (so that the Bag of Holding entry will turn
red if the player places too much inside it).