Fix some spell prerequisites
We need to be careful when the prereq list is "spell s (which is in
college c) plus n other spells from college c".  In that case, the
count needs to be n+1, not n, because spell s is in that college,
and gcs does not have special logic to avoid counting it.

Alter Terrain
Choke
Ethereal Body
False Memory
Great Geas
Lesser Geas
Paralyze Limb