Fix for #241
Behavior should be better now (it is in all of my test cases, but I
don't have the file for the specific one reported in issue #241). Things
still jump around a little bit between different scales that aren't
even multiples of the original size, but that is due to cumulative
floating-point errors and can't be reasonably addressed given the
current way I manage scaling.