Always set min & max sizes when enforcing window sizing
Unfortunately, the native window code doesn't call into the
normal resize method, so we have to explicitly set min/max in
order for the native code to respect our wishes. :-(

This has the downside that changes to the content that might
otherwise adjust min/max for a window will no longer be seen
after the first resize of the window. To address that, code
will need to explicitly set min/max on the window back to
null after changing the window content's min/max.