Initial Coding for true Phoenix d3 implementation
Damage progression, dice, weapondamage and weaponlist editor.
The code is meant to be as d3 agnostic as possible where it can be, so it doesn't actually specify d3 sides in addPercent function for example./
The addDice function is also changed to only assume d3s when Phoenix D3 is active, but it can take d6s or d3s as input since it converts to int to get d3s anyway.