Genericization
Added a check to prevent division by zero.
AddDice has different behaviours based on a setting in SheetSettings;
When dice sides are now equal, it has 1 of 3 behaviours: convert everything to the lower die size, the higher die size, or just add the die together (taking higher die)

There is a new heading for damage dice.
Die conversion is now a generic option that's not part of phoenixflame d3. It can be turned on or off and configured from any die size from d3 to d20.
Die addition is also configurable, it defaults to the existing default GCS behavior.
The percent bonus for swing is now a generic setting rather than part of phoenix d3

Minor Code Cleanup & some Final Changes

I don't know why that import statement was there
Ran the formatter on this.
Removed the damage progression specific code, it is now a general setting
Added the percent bonus to thrust too.