Fixed infinite loop caused by excessively large row data
I have to guard against the case where a single row is so large it
can't fit on a single page on its own. In this case, I just let it flow
off the end and drop that extra content.

In the future, see if I can do sub-row partitioning.