Fix #634: Relative paths within markdown files don't work initially
If you try to use a relative path from an unsaved markdown file, it
won't work because there is no path to be relative from yet. This is
expected behavior. However, if you then save it, those links should now
work, assuming the relative path does exist. This fixes the code to make
that in fact be the case.