Adjust encumbrance panel display
Equipment with the "Ignore weight for skills" option selected is no
longer also ignored for purposes of displaying the encumbrance penalty
for the weight being carried. I believe this inadvertently changed in
the update to v5.

Additionally, that option now only takes effect for the equipment if it
is also equipped.