Change the filter multi-select method
I've altered the code to require the shift or command/control key to be
pressed when selecting items from the tag filter if you want to use
multi-select. This way, the old behavior remains and is easy to use,
while those that need/want the multi-select behavior can also easily
add it when they actually need it.