Closes #658: Allow specifying specific TL for skill in template
If no TL is specified, but the checkbox to require a TL is checked, the
behavior is as before (i.e. when added to a character sheet, the
character's TL will be used). If, however, you do provide a TL, that TL
will be used instead, regardless of the character's TL.